# laravel-jwt-api-starter/docs/api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Marketplace API",
        "version": "0.0.1",
...
    },
    "servers": [
        {
            "url": "http://localhost/api"
        }
    ],
...
    "security": [
        {
            "http": []
        }
    ],
    "paths": {
...
        "/user/password": {
            "put": {
                "operationId": "profile.updatePassword",
                "description": "Change the authenticated user's password after validating the current password.",
                "summary": "Update password",
                "tags": [
                    "User/Profile"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Requests.UpdatePasswordRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`UserResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/Resources.UserResource"
                                        }
                                    },
                                    "required": [
                                        "user"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Attempting to access a route or perform an action without the required user role or permissions.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error_code": {
                                            "type": "string",
                                            "description": "The global error code.",
                                            "example": "FORBIDDEN"
                                        },
                                        "exception_type": {
                                            "type": "string",
                                            "description": "The exception class name.",
                                            "example": "AuthorizationException"
                                        },
                                        "message": {
                                            "type": "string",
                                            "description": "A human-readable error message.",
                                            "example": "You do not have permission to perform this action."
                                        }
                                    },
                                    "required": [
                                        "error_code",
                                        "exception_type",
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Requesting protected routes without a valid JWT token, or token is expired/invalid.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error_code": {
                                            "type": "string",
                                            "description": "The global error code.",
                                            "example": "UNAUTHENTICATED"
                                        },
                                        "exception_type": {
                                            "type": "string",
                                            "description": "The exception class name.",
                                            "example": "AuthenticationException"
                                        },
                                        "message": {
                                            "type": "string",
                                            "description": "A human-readable error message.",
                                            "example": "You are not authenticated. Please provide a valid token."
                                        }
                                    },
                                    "required": [
                                        "error_code",
                                        "exception_type",
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error_code": {
                                            "type": "string",
                                            "description": "The specific error code identifying a validation failure.",
                                            "example": "VALIDATION_ERROR"
                                        },
                                        "message": {
                                            "type": "string",
                                            "description": "Errors overview message.",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "description": "A detailed description of each field that failed validation.",
                                            "additionalProperties": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string",
                                                    "example": "The field is required."
                                                }
                                            }
                                        }
                                    },
                                    "required": [
                                        "error_code",
                                        "message",
                                        "errors"
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "http": {
                "type": "http",
                "scheme": "bearer"
            }
        },
        "schemas": {
...
            "Requests.UpdatePasswordRequest": {
                "type": "object",
                "properties": {
                    "current_password": {
                        "type": "string",
                        "description": "The user's current password."
                    },
                    "password": {
                        "type": "string",
                        "description": "The new password (must be confirmed and meet security requirements)."
                    },
                    "password_confirmation": {
                        "type": "string",
                        "description": "Confirmation of the new password."
                    }
                },
                "required": [
                    "current_password",
                    "password",
                    "password_confirmation"
                ],
                "title": "UpdatePasswordRequest"
            },
...
            "Resources.UserResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string"
                    },
                    "is_active": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "id",
                    "name",
                    "role",
                    "is_active"
                ],
                "title": "UserResource"
            }
        }
    }
}